Fix potential buffer overflow caused by sieve_init()
On systems where the size of size_t is less than that of uint64_t,
specifying a large value for the 'limit' parameter of sieve_init()
will result in an integer overflow of the size_t variable used
during allocation of the sieve's internal memory. This leads to a
buffer overflow when populating the sieve.